@@ -6,11 +6,9 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.globe.R
-import kotlinx.android.synthetic.main.community_fragment.*
+import com.globe.databinding.CommunityFragmentBinding
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 /**
@@ -22,14 +20,18 @@ class CommunityFragment : Fragment() {
 
     private val postListAdapter = PostListAdapter()
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.community_fragment, container, false)
+    private var _binding: CommunityFragmentBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
+        _binding = CommunityFragmentBinding.inflate(inflater, container, false)
+        return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        postListRecyclerView.run{
+        binding.postListRecyclerView.run {
             addItemDecoration(DividerItemDecoration(context, LinearLayoutManager.VERTICAL))
             layoutManager = LinearLayoutManager(context)
 
@@ -41,7 +43,7 @@ class CommunityFragment : Fragment() {
 
 
 
-        communityViewModel.postList.observe(viewLifecycleOwner, Observer {
+        communityViewModel.postList.observe(viewLifecycleOwner, {
             postListAdapter.items = it
         })
 
