@@ -2,32 +2,33 @@ package com.globe.rolly.ui.my_page
 
 import android.content.Intent
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.lifecycle.Observer
+import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.globe.rolly.ui.MainViewModel
-
 import com.globe.R
-import kotlinx.android.synthetic.main.my_page_home_fragment.*
-import timber.log.Timber
+import com.globe.databinding.MyPageHomeFragmentBinding
+import com.globe.rolly.ui.MainViewModel
 
 class MyPageHomeFragment : Fragment() {
 
     //    val mainViewModel: MainViewModel by sharedViewModel()
-    lateinit var mainViewModel: MainViewModel
+    private lateinit var mainViewModel: MainViewModel
 
     companion object {
         val instance = MyPageHomeFragment()
     }
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.my_page_home_fragment, container, false)
+    private var _binding: MyPageHomeFragmentBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
+        _binding = MyPageHomeFragmentBinding.inflate(inflater, container, false)
+        return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -43,11 +44,10 @@ class MyPageHomeFragment : Fragment() {
     }
 
     private fun addClickListener() {
-        logoutBtn.setOnClickListener {
+        binding.logoutBtn.setOnClickListener {
             mainViewModel.logout()
         }
-
-        mypage_profile_edit_button.setOnClickListener {
+        binding.mypageProfileEditButton.setOnClickListener {
             val intent = Intent(activity!!, ProfileEditActivity::class.java)
             intent.putExtra(ProfileEditActivity.EDIT_NAME, mainViewModel.userName.value)
             intent.putExtra(ProfileEditActivity.EDIT_NATIONCODE, mainViewModel.userNationCode)
@@ -61,7 +61,7 @@ class MyPageHomeFragment : Fragment() {
     }
 
     private fun initRecyclerView() {
-        mypageReservationList.run {
+        binding.mypageReservationList.run {
             addItemDecoration(DividerItemDecoration(context, LinearLayoutManager.HORIZONTAL))
             layoutManager = LinearLayoutManager(context, RecyclerView.HORIZONTAL, false)
             adapter = MyPageReservationAdapter(context)
@@ -74,22 +74,22 @@ class MyPageHomeFragment : Fragment() {
     }
 
     private fun observeEvents() {
-        mainViewModel.userName.observe(viewLifecycleOwner, Observer {
-            nickname.text = it
+        mainViewModel.userName.observe(viewLifecycleOwner, {
+            binding.nickname.text = it
         })
-        mainViewModel.reservations.observe(viewLifecycleOwner, Observer {
-            (mypageReservationList.adapter as MyPageReservationAdapter).run {
+        mainViewModel.reservations.observe(viewLifecycleOwner, {
+            (binding.mypageReservationList.adapter as MyPageReservationAdapter).run {
                 addItem(it)
                 notifyDataSetChanged()
             }
         })
-        mainViewModel.following.observe(viewLifecycleOwner, Observer {
+        mainViewModel.following.observe(viewLifecycleOwner, {
             val followingString = String.format(resources.getString(R.string.following), it)
-            following.text = followingString
+            binding.following.text = followingString
         })
-        mainViewModel.follower.observe(viewLifecycleOwner, Observer {
+        mainViewModel.follower.observe(viewLifecycleOwner, {
             val followerString = String.format(resources.getString(R.string.follower), it)
-            follower.text = followerString
+            binding.follower.text = followerString
         })
     }
 }
