@@ -23,8 +23,7 @@ import com.globe.rolly.network.model.user.signin.SignInModel
 import com.globe.rolly.network.model.user.signin.SignInRequestModel
 import com.globe.rolly.network.model.user.signup.SignUpRequestModel
 import com.globe.rolly.network.model.user.signup.SignUpResponseModel
-import io.reactivex.Observable
-import io.reactivex.Single
+import io.reactivex.rxjava3.core.Observable
 import okhttp3.ResponseBody
 import retrofit2.http.Body
 import retrofit2.http.GET
@@ -33,19 +32,19 @@ import retrofit2.http.POST
 interface RollyGlobeApiInterface {
 
     @GET("/codeNumbers/nationCode.json")
-    fun getNationCodeInfoList(): Single<List<NationCodeResponseModel>>
+    fun getNationCodeInfoList(): Observable<List<NationCodeResponseModel>>
 
     @POST("/ajax/user.php")
-    fun signUp(@Body param: SignUpRequestModel): Single<SignUpResponseModel>
+    fun signUp(@Body param: SignUpRequestModel): Observable<SignUpResponseModel>
 
     @POST("/ajax/user.php")
-    fun signIn(@Body param: SignInRequestModel): Single<SignInModel>
+    fun signIn(@Body param: SignInRequestModel): Observable<SignInModel>
 
     @POST("/ajax/spot.php")
-    fun getRecommendList(@Body param: RecommendRequestModel): Single<RecommendListResponseModel>
+    fun getRecommendList(@Body param: RecommendRequestModel): Observable<RecommendListResponseModel>
 
     @POST("/ajax/spot.php")
-    fun getSpotInnerContents(@Body param: InnerContentsRequestModel): Single<ResponseBody>
+    fun getSpotInnerContents(@Body param: InnerContentsRequestModel): Observable<ResponseBody>
 
     @POST("/ajax/spot.php")
     fun getGeocodeByGps(@Body param: GeocodeByGpsRequestModel): Observable<GeocodeByGpsResponseModel>
@@ -54,28 +53,28 @@ interface RollyGlobeApiInterface {
     fun getGeocodeByPlaceId(@Body param: GeocodeByPlaceIdRequestModel): Observable<GeocodeByPlaceIdResponseModel>
 
     @POST("ajax/mypage.php")
-    fun getMyPageHomeInfo(@Body param: MyPageHomeRequestModel): Single<MyPageHomeInfoResponseModel>
+    fun getMyPageHomeInfo(@Body param: MyPageHomeRequestModel): Observable<MyPageHomeInfoResponseModel>
 
     @POST("ajax/user.php")
-    fun editUserName(@Body param: EditUserNameRequestModel): Single<EditUserNameResponseModel>
+    fun editUserName(@Body param: EditUserNameRequestModel): Observable<EditUserNameResponseModel>
 
     @POST("ajax/user.php")
-    fun editUserPhoneNumber(@Body param: EditUserPhoneNumberRequestModel): Single<EditUserPhoneNumberResponseModel>
+    fun editUserPhoneNumber(@Body param: EditUserPhoneNumberRequestModel): Observable<EditUserPhoneNumberResponseModel>
 
     @POST("ajax/user.php")
-    fun editUserEmail(@Body param: EditUserEmailRequestModel): Single<EditUserEmailResponseModel>
+    fun editUserEmail(@Body param: EditUserEmailRequestModel): Observable<EditUserEmailResponseModel>
 
     @POST("ajax/user.php")
-    fun editUserGender(@Body param: EditUserGenderRequestModel): Single<EditUserGenderResponseModel>
+    fun editUserGender(@Body param: EditUserGenderRequestModel): Observable<EditUserGenderResponseModel>
 
     @POST("ajax/user.php")
-    fun editUserBirthday(@Body param: EditUserBirthdayRequestModel): Single<EditUserBirthdayResponseModel>
+    fun editUserBirthday(@Body param: EditUserBirthdayRequestModel): Observable<EditUserBirthdayResponseModel>
 
     @POST("ajax/user.php")
-    fun editUserPassword(@Body param: EditUserPasswordRequestModel): Single<EditUserPasswordResponseModel>
+    fun editUserPassword(@Body param: EditUserPasswordRequestModel): Observable<EditUserPasswordResponseModel>
 
     @POST("ajax/comment.php")
-    fun loadCommentList(@Body param: LoadCommentListRequestModel): Single<List<LoadCommentListResponseModel>>
+    fun loadCommentList(@Body param: LoadCommentListRequestModel): Observable<List<LoadCommentListResponseModel>>
 
     @POST("ajax/post.php")
     fun getPostList(@Body param : GetPostListRequestModel): Observable<GetPostListResponseModel>
