@@ -1,7 +1,7 @@
 package com.globe.rolly.support.baseclass
 
 import androidx.lifecycle.ViewModel
-import io.reactivex.disposables.CompositeDisposable
+import io.reactivex.rxjava3.disposables.CompositeDisposable
 import org.koin.core.KoinComponent
 
 open class BaseViewModel: ViewModel() , KoinComponent{
