@@ -1,19 +1,18 @@
 package com.globe.rolly.ui.community
 
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.globe.R
+import com.globe.databinding.PostListItemBinding
 import com.globe.rolly.network.model.response_model.PostInfo
 import com.globe.rolly.support.baseclass.BaseSimpleAdapter
-import kotlinx.android.synthetic.main.post_list_item.view.*
 
 class PostListAdapter : BaseSimpleAdapter<PostInfo, PostListAdapter.PostListViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): PostListViewHolder {
-        return PostListViewHolder(LayoutInflater.from(parent.context).inflate(R.layout.post_list_item, parent, false))
+        val binding = PostListItemBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+        return PostListViewHolder(binding)
     }
 
     override fun onBindViewHolder(holder: PostListViewHolder, position: Int) {
@@ -21,34 +20,34 @@ class PostListAdapter : BaseSimpleAdapter<PostInfo, PostListAdapter.PostListView
     }
 
 
-    inner class PostListViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+    inner class PostListViewHolder(private val itemBinding: PostListItemBinding) : RecyclerView.ViewHolder(itemBinding.root) {
 
 
         fun bind(post: PostInfo) {
-            itemView.userNickname.text = post.userNickName
+            itemBinding.userNickname.text = post.userNickName
 
             bindPostRegDate(post.postRegdate)
 
             Glide.with(itemView.context)
                 .load("https://test.rollyglobe.com/post/pics/small/" + post.postPicInfo[0].pictureRegDate + "/" + post.postPicInfo[0].postPicName)
-                .into(itemView.postPictures)
+                .into(itemBinding.postPictures)
 
-            itemView.postContent.text = post.postContent + "임시 콘텐츠"
+            itemBinding.postContent.text = post.postContent + "임시 콘텐츠"
 
-            itemView.spotPosition.text = "${post.continent} - ${post.nation} - ${post.city}"
+            itemBinding.spotPosition.text = "${post.continent} - ${post.nation} - ${post.city}"
 
-            itemView.spotName.text = post.spotTitleKor
+            itemBinding.spotName.text = post.spotTitleKor
 
-            itemView.spotIntro.text = post.spotIntro
+            itemBinding.spotIntro.text = post.spotIntro
 
-            itemView.likeCnt.text = "좋아요 77개"
-            itemView.commentCnt.text = "댓글 32개"
+            itemBinding.likeCnt.text = "좋아요 77개"
+            itemBinding.commentCnt.text = "댓글 32개"
         }
 
         private fun bindPostRegDate(regDate: String) {
             val refined = regDate.substring(0, 4) + "." + regDate.substring(4, 6) + "." + regDate.substring(6, 8)
 
-            itemView.postRegDate.text = refined
+            itemBinding.postRegDate.text = refined
         }
 
     }
