@@ -3,27 +3,24 @@ package com.globe.rolly.ui.community.writepost
 import android.app.Activity
 import android.util.DisplayMetrics
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import androidx.core.view.marginRight
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.globe.R
+import com.globe.databinding.ImageListItemBinding
 import com.globe.rolly.support.baseclass.BaseSimpleAdapter
-import kotlinx.android.synthetic.main.image_list_item.view.*
-import timber.log.Timber
 
 class ImageListAdapter : BaseSimpleAdapter<String, ImageListAdapter.ImageListViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ImageListViewHolder {
-        val holder = ImageListViewHolder(LayoutInflater.from(parent.context).inflate(R.layout.image_list_item, parent, false))
+        val binding = ImageListItemBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+        val holder = ImageListViewHolder(binding)
 
         val metrics = DisplayMetrics()
         (holder.itemView.context as Activity).windowManager.defaultDisplay.getMetrics(metrics)
         val width = metrics.widthPixels
 
-        holder.itemView.layoutParams.width = width/3 - 2
-        holder.itemView.layoutParams.height = width/3 - 4
+        holder.itemView.layoutParams.width = width / 3 - 2
+        holder.itemView.layoutParams.height = width / 3 - 4
         holder.itemView.requestLayout()
 
         return holder
@@ -34,9 +31,9 @@ class ImageListAdapter : BaseSimpleAdapter<String, ImageListAdapter.ImageListVie
         holder.bind(items[position])
     }
 
-    inner class ImageListViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+    inner class ImageListViewHolder(private val itemBinding: ImageListItemBinding) : RecyclerView.ViewHolder(itemBinding.root) {
         fun bind(imgUri: String) {
-            Glide.with(itemView.context).load(imgUri).into(itemView.img)
+            Glide.with(itemView.context).load(imgUri).into(itemBinding.img)
         }
 
     }
